+++ packages/gamut-labs/src/landingPage/Feature.tsx
@@ -36,38 +36,60 @@ const FeatureBlock = styled.div`
   }
 `;
 
-type FeaturedMediaProps = {
-  featuresMedia?: 'image' | 'icon' | 'stat' | 'none';
+type FeaturedImageProps = {
+  featuresMedia: 'image';
   imgSrc: string;
-  imgAlt?: string;
-  statText?: string;
+  imgAlt: string;
 };
 
-const FeaturedMedia: React.FC<FeaturedMediaProps> = ({
-  featuresMedia = 'none',
-  ...rest
-}) => {
-  if (featuresMedia === 'image') {
+type FeaturedIconProps = {
+  featuresMedia: 'icon';
+  imgSrc: string;
+  imgAlt: string;
+};
+
+type FeaturedStatProps = {
+  featuresMedia: 'stat';
+  statText: string;
+};
+
+type FeaturedNoMediaProps = {
+  featuresMedia: 'none';
+};
+
+type FeaturedMediaProps =
+  | FeaturedImageProps
+  | FeaturedIconProps
+  | FeaturedStatProps
+  | FeaturedNoMediaProps;
+
+const FeaturedMedia: React.FC<FeaturedMediaProps> = (props) => {
+  if (props.featuresMedia === 'image') {
     return (
-      <Image src={rest.imgSrc} alt={rest.imgAlt} data-testid="feature-image" />
+      <Image
+        src={props.imgSrc}
+        alt={props.imgAlt}
+        data-testid="feature-image"
+      />
     );
   }
 
-  if (featuresMedia === 'icon') {
+  if (props.featuresMedia === 'icon') {
     return (
-      <Icon src={rest.imgSrc} alt={rest.imgAlt} data-testid="feature-icon" />
+      <Icon src={props.imgSrc} alt={props.imgAlt} data-testid="feature-icon" />
     );
   }
 
-  if (featuresMedia === 'stat') {
+  if (props.featuresMedia === 'stat') {
     return (
       <Text
         as="div"
         marginTop={48}
         fontSize={{ xs: 44, lg: 64 }}
+        fontWeight="title"
         data-testid="feature-stat"
       >
-        {rest.statText}
+        {props.statText}
       </Text>
     );
   }
@@ -82,24 +104,16 @@ export type FeatureProps = Pick<
   FeaturedMediaProps;
 
 export const Feature: React.FC<FeatureProps> = ({
-  featuresMedia,
-  imgSrc,
-  imgAlt = '',
-  statText,
   title,
   desc,
   onAnchorClick,
   testId,
+  ...featuredMediaProps
 }) => (
   <FeatureBlock data-testid={testId}>
-    <FeaturedMedia
-      featuresMedia={featuresMedia}
-      imgSrc={imgSrc}
-      imgAlt={imgAlt}
-      statText={statText}
-    />
+    <FeaturedMedia {...featuredMediaProps} />
     {title && (
-      <Text as="h3" fontSize={{ xs: 22, lg: 26 }}>
+      <Text as="h3" fontSize={{ xs: 22, lg: 26 }} fontWeight="title">
         {title}
       </Text>
     )}